Split column item buttons into their own folders
