+++ ui/src/Common/ColumnItemButtons/CancelButton/CancelButton.tsx
@@ -1,0 +1,38 @@
+/*
+ * Copyright (c) 2022 Ford Motor Company
+ * All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import React, { ComponentPropsWithoutRef, forwardRef, LegacyRef } from 'react';
+import classnames from 'classnames';
+
+const CancelButton = forwardRef(
+	(
+		props: ComponentPropsWithoutRef<'button'>,
+		ref: LegacyRef<HTMLButtonElement>
+	) => {
+		const { className, children, ...buttonProps } = props;
+		return (
+			<button
+				className={classnames('column-item-button button-secondary', className)}
+				{...buttonProps}
+				ref={ref}
+			>
+				{children}
+			</button>
+		);
+	}
+);
+
+export default CancelButton;
