+++ ui/src/Common/ColumnItemButtons/CheckboxButton/CheckboxButton.scss
@@ -1,0 +1,45 @@
+/*
+ * Copyright (c) 2022 Ford Motor Company
+ * All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+.column-item-button {
+	.checkbox {
+		position: relative;
+
+		.checkbox-border {
+			$size: 15px;
+			width: $size;
+			height: $size;
+
+			background-color: transparent;
+			border: 2px solid;
+			border-radius: 4px;
+		}
+
+		.fa-check {
+			font-size: 1.1rem;
+
+			position: absolute;
+			top: -1px;
+			left: -1px;
+		}
+	}
+
+	&:not(.disabled) {
+		.checked .checkbox-border {
+			opacity: 0.25;
+		}
+	}
+}
