+++ ui/src/Common/ColumnItemButtons/ColumnItemButtonGroup/ColumnItemButtonGroup.scss
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-@use '../../Styles/main';
+@use '../../../Styles/main';
 
 $opacity: 0.25;
 
@@ -39,32 +39,30 @@ $opacity: 0.25;
 			border-bottom-right-radius: inherit;
 		}
 	}
-}
 
-.column-item-button {
-	display: flex;
-	flex: 25%;
-	align-items: center;
-	justify-content: center;
+	.column-item-button {
+		display: flex;
+		flex: 25%;
+		align-items: center;
+		justify-content: center;
 
-	color: inherit;
-	text-align: center;
+		color: inherit;
+		text-align: center;
 
-	background-color: transparent;
-	border: 0;
-	box-shadow: none;
-	cursor: pointer;
+		background-color: transparent;
+		border: 0;
+		box-shadow: none;
+		cursor: pointer;
 
-	user-select: none;
+		user-select: none;
 
-	position: relative;
+		position: relative;
 
-	&:not(:last-child) {
-		border-right: 2px solid rgba(main.$asphalt, 0.06);
-	}
+		&:not(:last-child) {
+			border-right: 2px solid rgba(main.$asphalt, 0.06);
+		}
 
-	&:not(.readonly) {
-		&[disabled],
+		&:disabled,
 		&.disabled {
 			user-select: none;
 			pointer-events: none;
@@ -73,44 +71,25 @@ $opacity: 0.25;
 				opacity: $opacity;
 			}
 		}
-	}
 
-	&.readonly {
-		user-select: none;
-		pointer-events: none;
-	}
-
-	.checkbox {
-		position: relative;
-
-		.checkbox-border {
-			$size: 15px;
-			width: $size;
-			height: $size;
-
-			background-color: transparent;
-			border: 2px solid;
-			border-radius: 4px;
+		&.button-primary,
+		&.button-secondary {
+			font-size: 1rem;
 		}
 
-		.fa-check {
-			font-size: 1.1rem;
-
-			position: absolute;
-			top: -1px;
-			left: -1px;
+		&.button-secondary {
+			&:hover,
+			&:focus {
+				color: main.$purple;
+			}
 		}
-	}
 
-	&:not(.disabled) {
-		.checked .checkbox-border {
-			opacity: $opacity;
+		&.button-primary {
+			&:hover,
+			&:focus {
+				background-color: main.$purple;
+			}
 		}
-	}
-
-	&.button-primary,
-	&.button-secondary {
-		font-size: 1rem;
 	}
 }
 
@@ -123,6 +102,13 @@ $opacity: 0.25;
 
 			background-color: inherit;
 			border-color: main.$task-footer-separator-dark;
+
+			&.button-secondary {
+				&:hover,
+				&:focus {
+					color: main.$dark-gray;
+				}
+			}
 
 			&.button-primary {
 				&:hover,
