+++ ui/src/Common/ColumnItemButtons/ConfirmButton/ConfirmButton.spec.tsx
@@ -1,0 +1,39 @@
+/*
+ * Copyright (c) 2022 Ford Motor Company
+ * All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import React from 'react';
+import { render, screen } from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
+
+import ConfirmButton from './ConfirmButton';
+
+describe('ConfirmButton', () => {
+	const mockConfirm = jest.fn();
+
+	it('should display confirm text', () => {
+		render(<ConfirmButton>Confirm</ConfirmButton>);
+
+		expect(screen.getByText('Confirm')).toBeDefined();
+	});
+
+	it('should handle clicks', () => {
+		render(<ConfirmButton onClick={mockConfirm}>Confirm</ConfirmButton>);
+
+		userEvent.click(screen.getByText('Confirm'));
+
+		expect(mockConfirm).toHaveBeenCalledTimes(1);
+	});
+});
