+++ ui/src/Common/ColumnItemButtons/DeleteButton/DeleteButton.tsx
@@ -1,0 +1,36 @@
+/*
+ * Copyright (c) 2022 Ford Motor Company
+ * All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import React, { ComponentPropsWithoutRef } from 'react';
+
+import Tooltip from '../../Tooltip/Tooltip';
+
+const DeleteButton = (props: ComponentPropsWithoutRef<'button'>) => {
+	const { ...buttonProps } = props;
+
+	return (
+		<button
+			{...buttonProps}
+			className="column-item-button"
+			data-testid="deleteButton"
+		>
+			<i className="fas fa-trash" aria-hidden="true" data-testid="trashIcon" />
+			<Tooltip>Delete</Tooltip>
+		</button>
+	);
+};
+
+export default DeleteButton;
