+++ ui/src/Common/ColumnItemButtons/EditButton/EditButton.spec.tsx
@@ -1,0 +1,56 @@
+/*
+ * Copyright (c) 2022 Ford Motor Company
+ * All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import * as React from 'react';
+import { render, screen } from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
+
+import EditButton from './EditButton';
+
+describe('EditButton', () => {
+	const mockEdit = jest.fn();
+
+	it('should display edit icon', () => {
+		render(<EditButton />);
+
+		screen.getByTestId('editIcon');
+	});
+
+	it('should have tooltip', () => {
+		render(<EditButton />);
+
+		screen.getByText('Edit');
+	});
+
+	it('should handle clicks', () => {
+		render(<EditButton onClick={mockEdit} />);
+
+		userEvent.click(screen.getByTestId('editIcon'));
+
+		expect(mockEdit).toHaveBeenCalledTimes(1);
+	});
+
+	it('can be disabled', () => {
+		render(<EditButton onClick={mockEdit} disabled={true} />);
+		const editButton = screen.getByTestId('editButton');
+
+		userEvent.click(editButton);
+
+		expect(mockEdit).toHaveBeenCalledTimes(0);
+		expect(editButton.getAttribute('disabled')).not.toBeNull();
+	});
+});
